Move all filter utils in Utils/filters
